Refactor cart service: implemented business logic, improved entity relationships, added OpenAPI docs, enhanced JWT handling, and integrated catalog client. Removed docker-compose dependency.
Took 4 hours 16 minutes

+++ cart-service/src/main/java/org/booknest/cartservice/repo/CartRepository.java
@@ -1,6 +1,7 @@
 package org.booknest.cartservice.repo;
 
 import org.booknest.cartservice.entity.Cart;
+import org.booknest.cartservice.entity.CartItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -10,7 +11,7 @@ import java.util.Optional;
 public interface CartRepository extends JpaRepository<Cart,Integer> {
 
     Optional<Cart> findByCartId(int cartId);
-    Optional<Cart> findByUserId(int userId);
-    boolean existsByUserId(int userId);
-    void deleteByUserId(int userId);
+    Optional<Cart> findByUserId(Long userId);
+    boolean existsByUserId(Long userId);
+    void deleteByUserId(Long userId);
 }
